msm: mdss: improve performance coming out of idle pc
When coming out of idle pc, all hw register need to be reprogrammed this
include huge amount of LUT and others. In order to speed this up,
register access clocks are increased but currently only during LUT
programming. Extend this for all other programming early so that the
commit is done faster and it doesn't miss the vsync window.

Bug: 77335499
Change-Id: I2519acb82570dcf4192807b0cdc26723bb6608f1
Signed-off-by: Adrian Salido <salidoa@google.com>
Signed-off-by: Lau <laststandrighthere@gmail.com>
Signed-off-by: Rsyd <rasyid2007alkautsar@gmail.com>